document production builds and catch-all
These two sections in the readme would have been very useful to me, hopefully others find it useful. It was not obvious to me production builds would not log or throw until I looked through the source code.  While looking at the source code, I found tests for throwOnUnhandled which I found extremely useful. I work on a large team, and not everyone is deeply experienced in Ember. There is nothing preventing warnings from finding their way into the logs.  Someone on the team referenced an old stack overflow answer which introduced a deprecation. Now that we have zero deprecation warnings, we want to keep it that way by forcing tests to fail.