Clean up test config for clarity
Remove a test for "modern" arguments which are now just used through the
whole codebase. Cause the test suite to throw when an unhandled
exception is encountered.

Remove all use RAISE_ON_DEPRECATION stuff, as this addon has its own
configuration for throwing on unhandled deprecation which should be
under test here.