feat: flushDeprecations merges detected deprecations with existing config
Currently calling flushDeprecations() is destructive in that it ignores any existing config. Users would need to merge the detected deprecations with their existing config manually. Which is error prone, and does not scale for very large monorepos.

This change is doing the merge automatically. Only minor drawback is that we need to serialize the merged config with `JSON.serialize`, so it is getting returned as e.g. `{ "handler": "silence", "matchId": "test" }` instead of `{ handler: 'silence', matchId: 'test' }`. But when using prettier, this will get auto-fixed anyway, so I think the trade-offs are positive here.